refactor: 调整项目名称校验规则 #147
refactor: 调整项目名称校验规则 #147

+++ src/backend/common/common-api/src/main/kotlin/com/tencent/bkrepo/common/api/util/HumanReadable.kt
@@ -20,8 +20,8 @@ object HumanReadable {
     }
 
     fun throughput(bytes: Long, nano: Long): String {
-        val speed = bytes / nano * 1000 * 1000 * 1000
-        return size(speed) + "/s"
+        val speed = bytes.toDouble() / nano * 1000 * 1000 * 1000
+        return size(speed.toLong()) + "/s"
     }
 
     fun time(nano: Long): String {
